feat: move mock to shared
