+++ packages/shared/src/index.ts
@@ -1,4 +1,5 @@
 export * from './math'
+export * from './mock/data-table'
 export * from './pzero'
 export * from './types/index'
 export * from './utils/functionHeader'
